+++ src/extension.ts
@@ -2,6 +2,7 @@ import {window, commands, ExtensionContext, workspace, Uri, Disposable, TreeData
 import {AnchorEngine} from './anchorEngine';
 import { FileAnchorProvider } from './fileAnchorProvider';
 import { WorkspaceAnchorProvider } from './workspaceAnchorProvider';
+import openAnchorList from './anchorListView';
 
 let anchorEngine: AnchorEngine;
 
@@ -19,6 +20,7 @@ export function activate(context: ExtensionContext) {
 	commands.registerCommand("commentAnchors.toggle", toggleVisibilitySetting);
 	commands.registerCommand("commentAnchors.openFileAndRevealLine", openFileAndRevealLine);
 	commands.registerCommand("commentAnchors.launchWorkspaceScan", launchWorkspaceScan);
+	commands.registerCommand("commentAnchors.listTags", () => openAnchorList(engine));
 
 	// Store a reference to the engine
 	anchorEngine = engine;